*) Code from Tiv - Added: Experimental code for support of the new emblems.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.x/GameServer@241 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ packets/PacketLib168.cs
@@ -770,7 +770,7 @@ namespace DOL.GS.PacketHandler
 			GSTCPPacketOut pak = new GSTCPPacketOut(GetPacketCode(ePackets.ObjectCreate));
 			pak.WriteShort((ushort)obj.ObjectID);
 			if (obj is GameStaticItem)
-				pak.WriteShort((obj as GameStaticItem).Emblem);
+				pak.WriteShort((ushort)(obj as GameStaticItem).Emblem);
 			else pak.WriteShort(0);
 			pak.WriteShort(obj.Heading);
 			pak.WriteShort((ushort)obj.Z);
